+++ lib/pkg/hid/types.go
@@ -1,0 +1,8 @@
+package hid
+
+// DeviceInfo represents keyboard handle information
+type DeviceInfo struct {
+	Name     string
+	Firmware string
+	Path     string
+}